Drawing is wrong?

+++ src/constants/types.rs
@@ -1,2 +1,2 @@
 pub type AreaId = u64;
-pub type BoidId = u64;
+pub type BoidId = usize;
